.tmux.conf, .gitignore, Makefile: License headers
... added just in case.  BlueOak seems to not require polluting files in a licensed project/repository with such headers:
https://writing.kemitchell.com/2020/12/27/War-on-License-Notices#:~:text=no%20copyright%20line,or%20special%20files

--- quote ---
no copyright line to fill out in the Blue Oak license, and no requirement to reproduce copyright notices in source code or special files
--- quote ---

But because this repo is more like a collection of autonomous script files (and even data files) which have the potential of being distributed and used separately, to me it looks worth to at least mention the license ID.  I chose SPDX format for that.

I also added my ID in case someone gets confused by the missing copyright text or just would like to contact me in future.  SPDX makes it look compact and easy to do.